proj2/requirements.txt
python-dotenv
requests
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-speech